📁 (#51) - move style file to under profile
