+++ nutella/src/components/Profile/Mypage/Contents/Profile.tsx
@@ -1,4 +1,4 @@
-import { ContentInner } from "../styles";
+import { ContentInner } from "../../styles";
 import * as S from "./styles";
 import { ArrowBlackIcons, GithubBlackIcons } from "../../../../assets/icons";
 import MarkdownRender from "../../../MarkdownRender";
